refactor: split region map markers into units and settlements parts (#3878)
Separate unit/selection marker paint from capital, town, and warp-zone
rendering so no single region_map marker part exceeds the Phase 3 size target.
